move utility functions into subfolder
